refactor(queue): extract exact review read model
